Split connected components into separate modules, fix SeedArtistForm unit test
